requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
opencv-python>=4.7.0
gymnasium>=0.29.0
matplotlib>=3.7.0
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0
Pillow>=9.5.0
scipy>=1.10.0
